Changed on Hand Quantity from double to decimal for Ingredients.

+++ MIS442_Final Project_RFernandez/BitsEFClasses/Models/Ingredient.cs
@@ -13,7 +13,7 @@ public partial class Ingredient
 
     public int IngredientTypeId { get; set; }
 
-    public double OnHandQuantity { get; set; }
+    public decimal OnHandQuantity { get; set; }
 
     public int UnitTypeId { get; set; }
